[feat] 주문 화면 조회 서버 연결

+++ app/src/main/java/com/codepatissier/keki/src/main/consumer/store/order/ConsumerOrderRetrofitInterface.kt
@@ -8,6 +8,6 @@ import retrofit2.http.Path
 interface ConsumerOrderRetrofitInterface {
     @GET("/orders/view/{storeIdx}")
     fun getConsumerOrderScreen(
-        @Path("storeIdx") storeIdx: Int
+        @Path("storeIdx") storeIdx: Long
     ): Call<ConsumerGetOrderScreenResponse>
 }
